backend/requirements.txt
fastapi
uvicorn[standard]
PyJWT
sqlmodel
httpx
prometheus-client
xmltodict
python-dotenv
passlib[bcrypt]
